+++ nms/src/cuda/nms_kernel.h
@@ -6,10 +6,10 @@ extern "C" {
 #endif
 
 #define DIVUP(m,n) ((m) / (n) + ((m) % (n) > 0))
-int const threadsPerBlock = sizeof(long) * 8;
+int const threadsPerBlock = sizeof(long long) * 8;
 
 void _nms(int boxes_num, float * boxes_dev,
-          long * mask_dev, float nms_overlap_thresh);
+          long long * mask_dev, float nms_overlap_thresh);
 
 #ifdef __cplusplus
 }
